requirements.txt
boto==2.48.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
DateTime==4.2
dill==0.2.7.1
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.1
google-auth==1.2.0
google-cloud==0.28.0
google-cloud-bigquery==0.28.0
google-cloud-bigtable==0.28.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==0.30.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.29.0
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.16.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==0.28.0
google-cloud-vision==0.28.0
google-compute-engine==2.7.5
google-gax==0.15.15
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.7.0
httplib2==0.10.3
idna==2.6
jdcal==1.3
numpy==1.13.3
oauth2client==3.0.0
openpyxl==2.4.9
pandas==0.21.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.4.0
psutil==5.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pyreadline==2.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
selenium==3.6.0
six==1.11.0
urllib3==1.22
zope.interface==4.4.3
